requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
aiohttp==3.9.3
google-generativeai==0.3.2
pandas==2.2.1
plotly==5.19.0
Pillow>=10.0.0
exifread>=3.0.0
asyncio>=3.4.3 